Php 8 + Laravel 9 (#29)
* Update npm packages

* Update composer packages

* Move lang folder, fix tlint

* Updated according laravel skeleton

* Fixes for laravel 9

* Change github workflow, lint

* Fix github workflow versions

+++ app/Services/Image/ImageService.php
@@ -26,7 +26,14 @@ class ImageService implements ImageServiceContract
 
     public function delete($path)
     {
-        return Storage::delete($path);
+        $exists = false;
+
+        if ($path != null) {
+            $exists = Storage::exists($path);
+            Storage::delete($path);
+        }
+
+        return $exists;
     }
 
     protected function stripImage(ImageInstance $img)
